requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.5
djangorestframework==3.14.0
fonttools==4.42.1
graphviz==0.20.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.25.2
packaging==23.1
pandas==2.1.0
patsy==0.5.3
Pillow==10.0.0
psycopg2-binary==2.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.2.0
tzdata==2023.3
xgboost==2.0.2
tpot
torch